delte order

+++ src/main/java/com/group2/glamping/model/entity/BookingSelection.java
@@ -31,5 +31,5 @@ public class BookingSelection {
     private String name;
 
     @Column(name = "quantity")
-    private double quantity;
+    private Integer quantity;
 }
